chatFeed is simplified the logic is in ChatBox
