+++ frontend/src/components/chat/chatBox/ChatFeed.tsx
@@ -1,52 +1,24 @@
-import { useEffect, useState, useRef } from 'react'
+import { useEffect, useRef } from 'react'
 import Msg from './Msg'
 import styles from './ChatFeed.module.css'
-import { useAppSelector } from '../../../store/types'
 
 interface ChatFeedProps {
-    socket: any
+    messages: any[]
 }
 
-const ChatFeed = ({ socket }: ChatFeedProps) => {
+const ChatFeed = ({ messages }: ChatFeedProps) => {
     const isFeedFull = useRef<HTMLDivElement>(null)
-    const currentChatSelected = useAppSelector(
-        (state) => state.chat.currentChatSelected
-    ) as number
-
-    socket.on('incomingMessage', (newMessage: any) => {
-        const msgCpy = [...msgs]
-        msgCpy.push(newMessage)
-        setMsgs(msgCpy)
-    })
-
-    const [msgs, setMsgs] = useState<any[]>([])
 
     useEffect(() => {
-        if (currentChatSelected) getAllMsgSocket()
-        else setMsgs([])
-    }, [currentChatSelected])
-
-    useEffect(() => {
-        isFeedFull.current?.scrollIntoView()
-    }, [msgs])
-
-    const getAllMsgSocket = () => {
-        socket.emit(
-            'findAllMsgByChannel',
-            currentChatSelected,
-            (response: any) => {
-                console.log(response)
-                setMsgs(response)
-            }
-        )
-    }
+        if (isFeedFull.current)
+            isFeedFull.current.scrollTop = isFeedFull.current?.scrollHeight
+    }, [messages])
 
     return (
-        <div className={styles.container}>
-            {msgs.map((msg) => (
+        <div className={styles.container} ref={isFeedFull}>
+            {messages.map((msg) => (
                 <Msg key={msg.id} msg={msg}></Msg>
             ))}
-            <div ref={isFeedFull}></div>
         </div>
     )
 }
